# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoSEO API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  pending: "pending",
  analyzing: "analyzing",
  active: "active",
  paused: "paused",
} as const;

export interface Project {
  id: number;
  name: string;
  websiteUrl: string;
  description?: string | null;
  niche?: string | null;
  targetAudience?: string | null;
  language: string;
  status: ProjectStatus;
  wordpressUrl?: string | null;
  wordpressUsername?: string | null;
  wordpressConnected?: boolean | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  websiteUrl: string;
  language?: string;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  pending: "pending",
  analyzing: "analyzing",
  active: "active",
  paused: "paused",
} as const;

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  niche?: string;
  targetAudience?: string;
  language?: string;
  status?: UpdateProjectBodyStatus;
}

export interface WebsiteAnalysis {
  niche: string;
  targetAudience: string;
  description: string;
  suggestedKeywords: string[];
}

export type KeywordSearchIntent =
  (typeof KeywordSearchIntent)[keyof typeof KeywordSearchIntent];

export const KeywordSearchIntent = {
  informational: "informational",
  navigational: "navigational",
  transactional: "transactional",
  commercial: "commercial",
} as const;

export type KeywordDifficulty =
  (typeof KeywordDifficulty)[keyof typeof KeywordDifficulty];

export const KeywordDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Keyword {
  id: number;
  projectId: number;
  keyword: string;
  searchIntent: KeywordSearchIntent;
  difficulty: KeywordDifficulty;
  priority: number;
  used: boolean;
  createdAt: string;
}

export interface ContentPlanItem {
  id: number;
  projectId: number;
  dayNumber: number;
  title: string;
  keyword: string;
  searchIntent: string;
  articleId?: number | null;
  scheduledDate?: string | null;
  createdAt: string;
}

export type ArticleStatus = (typeof ArticleStatus)[keyof typeof ArticleStatus];

export const ArticleStatus = {
  draft: "draft",
  generating: "generating",
  published: "published",
  failed: "failed",
} as const;

export interface Article {
  id: number;
  projectId: number;
  title: string;
  keyword: string;
  content?: string | null;
  metaDescription?: string | null;
  wordCount?: number | null;
  status: ArticleStatus;
  language: string;
  searchIntent?: string | null;
  wordpressPostId?: number | null;
  wordpressPublishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateArticleBody {
  title: string;
  keyword: string;
  searchIntent?: string | null;
  language?: string;
}

export type UpdateArticleBodyStatus =
  (typeof UpdateArticleBodyStatus)[keyof typeof UpdateArticleBodyStatus];

export const UpdateArticleBodyStatus = {
  draft: "draft",
  generating: "generating",
  published: "published",
  failed: "failed",
} as const;

export interface UpdateArticleBody {
  title?: string;
  content?: string;
  metaDescription?: string;
  status?: UpdateArticleBodyStatus;
}

export interface ProjectAnalytics {
  totalArticles: number;
  publishedArticles: number;
  draftArticles: number;
  totalWordCount: number;
  keywordsCount: number;
  /** Percentage of content plan items with articles (0-100) */
  contentPlanProgress: number;
  recentArticles: Article[];
}

export interface AnalyticsSummary {
  totalProjects: number;
  totalArticles: number;
  totalWordCount: number;
  publishedArticles: number;
  activeProjects: number;
}

export interface ConnectWordPressBody {
  /** WordPress site URL (e.g. https://myblog.com) */
  wordpressUrl: string;
  wordpressUsername: string;
  /** WordPress Application Password */
  wordpressAppPassword: string;
}

export interface WordPressConnectionStatus {
  connected: boolean;
  siteTitle?: string | null;
  error?: string | null;
}

export interface PublishToWordPressBody {
  /** If set, post is scheduled as a future post on this date */
  scheduledDate?: string | null;
}

export interface ScheduleAllResult {
  scheduled: number;
  skipped: number;
  errors: number;
}

export type ListArticlesParams = {
  status?: ListArticlesStatus;
};

export type ListArticlesStatus =
  (typeof ListArticlesStatus)[keyof typeof ListArticlesStatus];

export const ListArticlesStatus = {
  draft: "draft",
  generating: "generating",
  published: "published",
  failed: "failed",
} as const;