# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoSEO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/keywords:
    get:
      operationId: listKeywords
      tags: [keywords]
      summary: List keywords for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: generateKeywords
      tags: [keywords]
      summary: Generate keyword suggestions using AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Generated keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"

  # ── Content Plans ─────────────────────────────────────────────────────
  /projects/{id}/content-plan:
    get:
      operationId: getContentPlan
      tags: [contentPlans]
      summary: Get the 30-day content plan for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content plan items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentPlanItem"
    post:
      operationId: generateContentPlan
      tags: [contentPlans]
      summary: Generate a 30-day content plan using AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Generated content plan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentPlanItem"

  # ── Articles ──────────────────────────────────────────────────────────
  /projects/{projectId}/articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, generating, published, failed]
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create and generate an SEO article using AI
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /projects/{projectId}/articles/{articleId}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get an article by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article (edit content, status)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleBody"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/articles/{articleId}/regenerate:
    post:
      operationId: regenerateArticle
      tags: [articles]
      summary: Regenerate article content using AI
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Regenerated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

  # ── WordPress Integration ─────────────────────────────────────────────
...
components:
  schemas:
...
    Keyword:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        keyword:
          type: string
        searchIntent:
          type: string
          enum: [informational, navigational, transactional, commercial]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        priority:
          type: integer
        used:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, projectId, keyword, searchIntent, difficulty, priority, used, createdAt]
    ContentPlanItem:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        dayNumber:
          type: integer
        title:
          type: string
        keyword:
          type: string
        searchIntent:
          type: string
        articleId:
          type: integer
          nullable: true
        scheduledDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, projectId, dayNumber, title, keyword, searchIntent, createdAt]
    Article:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        keyword:
          type: string
        content:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        wordCount:
          type: integer
          nullable: true
        status:
          type: string
          enum: [draft, generating, published, failed]
        language:
          type: string
        searchIntent:
          type: string
          nullable: true
        wordpressPostId:
          type: integer
          nullable: true
        wordpressPublishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, title, keyword, status, language, createdAt, updatedAt]
    CreateArticleBody:
      type: object
      properties:
        title:
          type: string
        keyword:
          type: string
        searchIntent:
          type: string
          nullable: true
        language:
          type: string
      required: [title, keyword]
    UpdateArticleBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        metaDescription:
          type: string
        status:
          type: string
          enum: [draft, generating, published, failed]
...