# lib/api-zod/src/generated/types/contentPlanItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoSEO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentPlanItem {
  id: number;
  projectId: number;
  dayNumber: number;
  title: string;
  keyword: string;
  searchIntent: string;
  articleId?: number | null;
  scheduledDate?: Date | null;
  createdAt: Date;
}