# lib/api-zod/src/generated/types/createArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoSEO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateArticleBody {
  title: string;
  keyword: string;
  searchIntent?: string | null;
  language?: string;
}